Add to OrcJIT test coverage
The existing test involves a callback into Haskell. This segfaults on
darwin for a reason I have yet to determine. In lieu of fixing this --
and to exercise things a bit more -- I have added two more tests.

One is a simple main-style function that does not call back into
Haskell. The other is an arithmetic function that computes a result from
some numeric arguments. Both of these tests work properly on darwin.